CLIENT-GAME: Put debugCollisionRect in client side
MINOR
